Swallow <delete> and <enter>, prevent data loss
Disable delete when not editing text in form fields to prevent browsers
from navigating back and losing all changes since the last save. Same
for enter on the mediaDrawer search input.

Because we're manipulating JSON with javascript, we don't get the handy
"Are you sure you want to {close this tab,leave this page}" dialogs.
Maybe there's a way to trigger those?